sql,streamingccl: remove cutover builtin
This has been replaced with SQL syntax and isn't required any longer.

Epic: none

Release note: None
